rename key.rs and move Subject and Validor
