+++ src/lib.rs
@@ -1,10 +1,12 @@
 pub mod id_token;
 pub mod jwt;
-pub mod key;
+pub mod key_method;
 pub mod provider;
 pub mod relying_party;
 pub mod request;
 pub mod response;
+pub mod subject;
+pub mod validator;
 
 pub use id_token::IdToken;
 pub use jwt::JsonWebToken;
@@ -12,6 +14,8 @@ pub use provider::Provider;
 pub use relying_party::RelyingParty;
 pub use request::SiopRequest;
 pub use response::SiopResponse;
+pub use subject::Subject;
+pub use validator::Validator;
 
 #[cfg(test)]
 pub mod test_utils;
