+++ src/test_utils.rs
@@ -1,4 +1,4 @@
-use crate::{provider::Subject, relying_party::Validator};
+use crate::{Subject, Validator};
 use anyhow::Result;
 use async_trait::async_trait;
 use derivative::{self, Derivative};
